+++ changelog.md
@@ -21,6 +21,8 @@
 *  Refactored `EwbDataFilePaths`: 
   * The `EwbDataFilePaths` class has been refactored into an interface to enhance flexibility and abstraction. 
   * A new class, `LocalEwbDataFilePaths`, has been introduced to specifically handle the resolution of database paths for the local file system.
+* `Switch.ratedCurrent` has been converted to a `double` (used to be an `integer`). Type safe languages will need to be updated to support floating point
+  arithmatic/syntax.
 
 ### New Features
 * A file named after the ID of an ingestion job is now created when running `MetricsDatabaseWriter.save()`. For this feature to take effect, a `modelPath` must